Updated to more recent portable stdint. The older version was incompatible with newer windows SDKs.
